Revert "Merge pull request #87320 from microsoft/misolori/codicon-consolidation"
This reverts commit 6c9b23edc4329b3cc6faa6b69d219f3ee1f5a8fc, reversing
changes made to 085abd5f1a2c30abb8a10bd00be9b72291ed478d.
