+++ src/vs/base/browser/ui/codiconLabel/codicon/codicon-animations.css
