+++ src/vs/base/browser/ui/codiconLabel/codicon/codicon.css
