+++ src/vs/base/common/codicon.ts
