+++ src/vs/base/common/codicons.ts
