+++ src/vs/code/electron-browser/issue/issueReporterMain.ts
@@ -8,7 +8,7 @@ import * as os from 'os';
 import * as browser from 'vs/base/browser/browser';
 import { $ } from 'vs/base/browser/dom';
 import { Button } from 'vs/base/browser/ui/button/button';
-import { CodiconLabel } from 'vs/base/parts/codicon/browser/codiconLabel';
+import { CodiconLabel } from 'vs/base/browser/ui/codiconLabel/codiconLabel';
 import * as collections from 'vs/base/common/collections';
 import { debounce } from 'vs/base/common/decorators';
 import { Disposable } from 'vs/base/common/lifecycle';
