+++ tslint.json
@@ -68,8 +68,7 @@
 				"target": "**/vs/base/common/**",
 				"restrictions": [
 					"vs/nls",
-					"**/vs/base/common/**",
-					"**/vs/base/parts/*/common/**"
+					"**/vs/base/common/**"
 				]
 			},
 			{
@@ -79,8 +78,7 @@
 					"sinon",
 					"vs/nls",
 					"**/vs/base/common/**",
-					"**/vs/base/test/common/**",
-					"**/vs/base/parts/*/common/**"
+					"**/vs/base/test/common/**"
 				]
 			},
 			{
@@ -88,16 +86,14 @@
 				"restrictions": [
 					"vs/nls",
 					"vs/css!./**/*",
-					"**/vs/base/{common,browser}/**",
-					"**/vs/base/parts/*/{common,browser}/**"
+					"**/vs/base/{common,browser}/**"
 				]
 			},
 			{
 				"target": "**/vs/base/node/**",
 				"restrictions": [
 					"vs/nls",
-					"**/vs/base/{common,node}/**",
-					"**/vs/base/parts/*/{common,node}/**",
+					"**/vs/base/{common,browser,node}/**",
 					"!path" // node modules (except path where we have our own impl)
 				]
 			},
@@ -109,7 +105,6 @@
 					"sinon",
 					"vs/nls",
 					"**/vs/base/{common,browser}/**",
-					"**/vs/base/parts/*/{common,browser}/**",
 					"**/vs/base/test/{common,browser}/**"
 				]
 			},
@@ -134,8 +129,8 @@
 				"target": "**/vs/base/parts/*/node/**",
 				"restrictions": [
 					"vs/nls",
-					"**/vs/base/{common,node}/**",
-					"**/vs/base/parts/*/{common,node}/**",
+					"**/vs/base/{common,browser,node}/**",
+					"**/vs/base/parts/*/{common,browser,node}/**",
 					"!path" // node modules (except path where we have our own impl)
 				]
 			},
@@ -153,8 +148,8 @@
 				"target": "**/vs/base/parts/*/electron-main/**",
 				"restrictions": [
 					"vs/nls",
-					"**/vs/base/{common,node,electron-main}/**",
-					"**/vs/base/parts/*/{common,node,electron-main}/**",
+					"**/vs/base/{common,browser,node,electron-main}/**",
+					"**/vs/base/parts/*/{common,browser,node,electron-main}/**",
 					"!path" // node modules (except path where we have our own impl)
 				]
 			},
@@ -192,9 +187,9 @@
 				"target": "**/vs/platform/*/node/**",
 				"restrictions": [
 					"vs/nls",
-					"**/vs/base/{common,node}/**",
-					"**/vs/base/parts/*/{common,node}/**",
-					"**/vs/platform/*/{common,node}/**",
+					"**/vs/base/{common,browser,node}/**",
+					"**/vs/base/parts/*/{common,browser,node}/**",
+					"**/vs/platform/*/{common,browser,node}/**",
 					"!path" // node modules (except path where we have our own impl)
 				]
 			},
@@ -213,9 +208,9 @@
 				"target": "**/vs/platform/*/electron-main/**",
 				"restrictions": [
 					"vs/nls",
-					"**/vs/base/{common,node,electron-main}/**",
-					"**/vs/base/parts/*/{common,node,electron-main}/**",
-					"**/vs/platform/*/{common,node,electron-main}/**",
+					"**/vs/base/{common,browser,node}/**",
+					"**/vs/base/parts/*/{common,browser,node,electron-browser}/**",
+					"**/vs/platform/*/{common,browser,node,electron-main}/**",
 					"!path" // node modules (except path where we have our own impl)
 				]
 			},
@@ -401,7 +396,6 @@
 					"**/vs/editor/common/**",
 					"**/vs/editor/contrib/*/common/**",
 					"**/vs/workbench/api/common/**",
-					"**/vs/base/parts/*/common/**",
 					"**/vs/workbench/common/**",
 					"**/vs/workbench/services/*/common/**",
 					"**/vs/workbench/contrib/*/common/**"
@@ -515,6 +509,37 @@
 				]
 			},
 			{
+				"target": "**/vs/workbench/contrib/files/common/**",
+				"restrictions": [
+					"vs/nls",
+					"**/vs/base/common/**",
+					"**/vs/base/parts/*/common/**",
+					"**/vs/platform/*/common/**",
+					"**/vs/editor/common/**",
+					"**/vs/editor/contrib/*/common/**",
+					"**/vs/workbench/common/**",
+					"**/vs/workbench/services/*/common/**",
+					"**/vs/workbench/contrib/files/common/**",
+					"assert"
+				]
+			},
+			{
+				"target": "**/vs/workbench/contrib/files/browser/**",
+				"restrictions": [
+					"vs/nls",
+					"vs/css!./**/*",
+					"**/vs/base/{common,browser}/**",
+					"**/vs/base/parts/*/{common,browser}/**",
+					"**/vs/platform/*/{common,browser}/**",
+					"**/vs/editor/{common,browser}/**",
+					"**/vs/editor/contrib/**", // editor/contrib is equivalent to /browser/ by convention
+					"**/vs/workbench/{common,browser}/**",
+					"**/vs/workbench/services/*/{common,browser}/**",
+					"**/vs/workbench/contrib/files/{common,browser}/**",
+					"assert"
+				]
+			},
+			{
 				"target": "**/vs/workbench/contrib/terminal/browser/**",
 				"restrictions": [
 					// xterm and its addons are strictly browser-only components
@@ -541,10 +566,10 @@
 				"target": "**/vs/code/node/**",
 				"restrictions": [
 					"vs/nls",
-					"**/vs/base/**/{common,node}/**",
-					"**/vs/base/parts/**/{common,node}/**",
-					"**/vs/platform/**/{common,node}/**",
-					"**/vs/code/**/{common,node}/**",
+					"**/vs/base/**/{common,browser,node}/**",
+					"**/vs/base/parts/**/{common,browser,node}/**",
+					"**/vs/platform/**/{common,browser,node}/**",
+					"**/vs/code/**/{common,browser,node}/**",
 					"!path" // node modules (except path where we have our own impl)
 				]
 			},
@@ -564,10 +589,10 @@
 				"target": "**/vs/code/electron-main/**",
 				"restrictions": [
 					"vs/nls",
-					"**/vs/base/**/{common,node,electron-main}/**",
-					"**/vs/base/parts/**/{common,node,electron-main}/**",
-					"**/vs/platform/**/{common,node,electron-main}/**",
-					"**/vs/code/**/{common,node,electron-main}/**",
+					"**/vs/base/**/{common,browser,node}/**",
+					"**/vs/base/parts/**/{common,browser,node,electron-main}/**",
+					"**/vs/platform/**/{common,browser,node,electron-main}/**",
+					"**/vs/code/**/{common,browser,node,electron-main}/**",
 					"!path" // node modules (except path where we have our own impl)
 				]
 			},
